refactor(dashboard): udpate types
