+++ apps/dashboard/src/types/Admin.ts
@@ -1,0 +1,6 @@
+export type Admin = {
+    address: string
+    id: string
+}
+
+export default Admin
